ДОбавил абстрактный класс для диагностик метаданных
